# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QDIA Export API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon: string;
  /** @nullable */
  product_count?: number | null;
}

export interface IncotermPrices {
  exw: number;
  fob: number;
  cfr: number;
  cif: number;
  currency?: string;
  unit?: string;
}

export type ProductExportStatus = typeof ProductExportStatus[keyof typeof ProductExportStatus];


export const ProductExportStatus = {
  draft: 'draft',
  pending: 'pending',
  published: 'published',
  suspended: 'suspended',
} as const;

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  /** @nullable */
  sku?: string | null;
  /** @nullable */
  image_url?: string | null;
  images?: string[];
  supplier_id: number;
  /** @nullable */
  supplier_name?: string | null;
  /** @nullable */
  supplier_location?: string | null;
  moq: number;
  moq_unit: string;
  port_depart: string;
  /** @nullable */
  origin_wilaya?: string | null;
  certifications?: string[];
  /** @nullable */
  packaging?: string | null;
  /** @nullable */
  processing?: string | null;
  export_status: ProductExportStatus;
  prices: IncotermPrices;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  review_count?: number | null;
  /** @nullable */
  orders_fulfilled?: number | null;
  target_markets?: string[];
  is_featured?: boolean;
}

export interface ProductInput {
  name: string;
  description?: string;
  category: string;
  sku?: string;
  moq: number;
  moq_unit: string;
  port_depart: string;
  origin_wilaya?: string;
  certifications?: string[];
  packaging?: string;
  processing?: string;
  prices: IncotermPrices;
  target_markets?: string[];
}

export interface ProductList {
  data: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface Supplier {
  id: number;
  company_name: string;
  country: string;
  wilaya: string;
  verified: boolean;
  verification_level?: number;
  platform_years: number;
  response_rate: number;
  transaction_count: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  avatar?: string | null;
}

export type RfqStatus = typeof RfqStatus[keyof typeof RfqStatus];


export const RfqStatus = {
  pending: 'pending',
  quoted: 'quoted',
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

export interface Rfq {
  id: number;
  product_name: string;
  quantity: number;
  quantity_unit: string;
  destination_country: string;
  requested_incoterm: string;
  /** @nullable */
  target_price?: number | null;
  /** @nullable */
  message?: string | null;
  status: RfqStatus;
  created_at: string;
}

export interface RfqInput {
  product_name: string;
  product_description?: string;
  quantity: number;
  quantity_unit: string;
  destination_country: string;
  requested_incoterm: string;
  target_price?: number;
  message?: string;
}

export interface DashboardStats {
  active_inquiries: number;
  inquiries_change_pct?: number;
  pending_rfqs: number;
  pending_rfqs_urgent?: boolean;
  total_export_value: number;
  export_value_period?: string;
  store_visits: number;
  store_visits_scope?: string;
  total_products: number;
  active_listings: number;
  ai_suggestions: number;
}

export interface RfqActivity {
  id: number;
  rfq_ref: string;
  product_name: string;
  buyer_city: string;
  buyer_country: string;
  buyer_flag: string;
  time_ago: string;
}

export interface ProductPerformance {
  category: string;
  views: number;
}

export type ListProductsParams = {
/**
 * @nullable
 */
category_id?: number | null;
/**
 * @nullable
 */
search?: string | null;
/**
 * @nullable
 */
incoterm?: string | null;
/**
 * @nullable
 */
page?: number | null;
/**
 * @nullable
 */
limit?: number | null;
};