# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QDIA Export API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  active_inquiries: number;
  inquiries_change_pct?: number;
  pending_rfqs: number;
  pending_rfqs_urgent?: boolean;
  total_export_value: number;
  export_value_period?: string;
  store_visits: number;
  store_visits_scope?: string;
  total_products: number;
  active_listings: number;
  ai_suggestions: number;
}