# lib/api-zod/src/generated/types/incotermPrices.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QDIA Export API
 * OpenAPI spec version: 0.1.0
 */

export interface IncotermPrices {
  exw: number;
  fob: number;
  cfr: number;
  cif: number;
  currency?: string;
  unit?: string;
}