# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QDIA Export API
 * OpenAPI spec version: 0.1.0
 */
import type { IncotermPrices } from './incotermPrices';
import type { ProductExportStatus } from './productExportStatus';

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  /** @nullable */
  sku?: string | null;
  /** @nullable */
  image_url?: string | null;
  images?: string[];
  supplier_id: number;
  /** @nullable */
  supplier_name?: string | null;
  /** @nullable */
  supplier_location?: string | null;
  moq: number;
  moq_unit: string;
  port_depart: string;
  /** @nullable */
  origin_wilaya?: string | null;
  certifications?: string[];
  /** @nullable */
  packaging?: string | null;
  /** @nullable */
  processing?: string | null;
  export_status: ProductExportStatus;
  prices: IncotermPrices;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  review_count?: number | null;
  /** @nullable */
  orders_fulfilled?: number | null;
  target_markets?: string[];
  is_featured?: boolean;
}