# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QDIA Export API
 * OpenAPI spec version: 0.1.0
 */
import type { IncotermPrices } from './incotermPrices';

export interface ProductInput {
  name: string;
  description?: string;
  category: string;
  sku?: string;
  moq: number;
  moq_unit: string;
  port_depart: string;
  origin_wilaya?: string;
  certifications?: string[];
  packaging?: string;
  processing?: string;
  prices: IncotermPrices;
  target_markets?: string[];
}