# lib/api-zod/src/generated/types/productList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QDIA Export API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface ProductList {
  data: Product[];
  total: number;
  page: number;
  limit: number;
}