# lib/api-zod/src/generated/types/productPerformance.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QDIA Export API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductPerformance {
  category: string;
  views: number;
}