# lib/api-zod/src/generated/types/rfqActivity.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QDIA Export API
 * OpenAPI spec version: 0.1.0
 */

export interface RfqActivity {
  id: number;
  rfq_ref: string;
  product_name: string;
  buyer_city: string;
  buyer_country: string;
  buyer_flag: string;
  time_ago: string;
}