# lib/api-zod/src/generated/types/supplier.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QDIA Export API
 * OpenAPI spec version: 0.1.0
 */

export interface Supplier {
  id: number;
  company_name: string;
  country: string;
  wilaya: string;
  verified: boolean;
  verification_level?: number;
  platform_years: number;
  response_rate: number;
  transaction_count: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  avatar?: string | null;
}